+++ framework/server.js
@@ -37,6 +37,7 @@ class Server {
 			wsServer: false,
       enable_precompressed_negotiation: false,
       enable_service_worker: false,
+      defer_scripts: false,
       request_timeout_ms: 30000,
       headers_timeout_ms: 35000,
       keep_alive_timeout_ms: 5000